build(meos): align the extended-family build flags to the MobilityDB/MEOS convention
The extended temporal-type families are selected at build time with the same
uppercase flag names and ON|OFF (also 1|0) values as the MobilityDB/MEOS CMake
build, with matching defaults: NPOINT is included by default; CBUFFER, POSE,
RGEO, and H3 are excluded unless their flag is ON|1 (RGEO requires POSE). Passing
-DCBUFFER=ON (or =1) includes the tcbuffer facade family; -DNPOINT=OFF drops the
network-point family. Each family's MeosOps* facade sources and its per-family
smoke test (MeosCbufferSmokeTest, MeosNpointSmokeTest, MeosPoseSmokeTest) are
dropped from the build when the family is excluded; the core and geo smoke
checks remain in MeosFacadeSmokeTest.
